refactor(story-engine): split monolithic Service into domain services
- Extract ProjectService for project CRUD and story lifecycle
- Extract StoryService for generation, streaming, and history
- Extract CharacterService for extraction, refinement, render, and upload
- Wire each router to its domain service via dedicated DI factories
- Remove god-object Service class and single get_service dependency
- No facade, no inter-service coupling; repos remain the shared read layer
