+++ backend/core/story_engine/service/__init__.py
@@ -1,5 +1,9 @@
-from .service import Service
+from .character_service import CharacterService
+from .project_service import ProjectService
+from .story_service import StoryService
 
 __all__ = [
-    "Service",
+    "CharacterService",
+    "ProjectService",
+    "StoryService",
 ]
